Deprecate Limit and LimitAll; group deprecated API in deprecated.go
LimitBy is now the canonical entry point, so deprecate Limit (use
LimitBy(n, t, keyFn, opts...), passing the key explicitly or Key("*") for a
single global bucket) and LimitAll (use LimitBy(n, t, Key("*"))). Bodies are
unchanged behavior-wise, so existing code keeps compiling.

Move the whole deprecated surface (Limit, LimitAll, LimitByIP, LimitByRealIP,
KeyByRealIP, WithKeyByRealIP) into deprecated.go to keep httprate.go focused on
the supported API. Each deprecated wrapper now delegates to LimitBy, mirroring
its documented migration. httprate has not reached a stable v1, so these may be
removed in a future major release.

Update README and _example to use LimitBy instead of the deprecated Limit.
