+++ httprate.go
@@ -8,10 +8,6 @@ import (
 	"time"
 )
 
-func Limit(requestLimit int, windowLength time.Duration, options ...Option) func(next http.Handler) http.Handler {
-	return NewRateLimiter(requestLimit, windowLength, options...).Handler
-}
-
 // LimitBy is the canonical entry point for rate-limiting by an explicit key.
 //
 // It is shorthand for Limit with keyFn installed as the rate-limit key. The
@@ -29,7 +25,7 @@ func Limit(requestLimit int, windowLength time.Duration, options ...Option) func
 //	r.Use(httprate.LimitBy(100, time.Minute,
 //		httprate.JoinKeys(httprate.KeyByIP, httprate.KeyByEndpoint)))
 func LimitBy(requestLimit int, windowLength time.Duration, keyFn KeyFunc, options ...Option) func(next http.Handler) http.Handler {
-	return Limit(requestLimit, windowLength, append([]Option{WithKeyFuncs(keyFn)}, options...)...)
+	return NewRateLimiter(requestLimit, windowLength, append([]Option{WithKeyFuncs(keyFn)}, options...)...).Handler
 }
 
 type KeyFunc func(r *http.Request) (string, error)
@@ -42,28 +38,6 @@ type ResponseHeaders struct {
 	Increment  string // Default: X-RateLimit-Increment
 	Reset      string // Default: X-RateLimit-Reset
 	RetryAfter string // Default: Retry-After
-}
-
-func LimitAll(requestLimit int, windowLength time.Duration) func(next http.Handler) http.Handler {
-	return Limit(requestLimit, windowLength)
-}
-
-// Deprecated: Use LimitBy(requestLimit, windowLength, KeyByIP) instead. This is
-// an ergonomic rename with identical behavior, not a security fix.
-func LimitByIP(requestLimit int, windowLength time.Duration) func(next http.Handler) http.Handler {
-	return Limit(requestLimit, windowLength, WithKeyFuncs(KeyByIP))
-}
-
-// Deprecated: LimitByRealIP is built on the spoofable KeyByRealIP and lets a
-// remote attacker forge the rate-limit key — see GHSA-9g5q-2w5x-hmxf,
-// GHSA-rjr7-jggh-pgcp, GHSA-3fxj-6jh8-hvhx for the equivalent flaw in chi's
-// middleware.RealIP. Install one of chi's middleware.ClientIPFrom* middlewares
-// (chi v5.3.0+) and use LimitBy with KeyFromContext instead:
-//
-//	r.Use(middleware.ClientIPFromXFF("10.0.0.0/8"))
-//	r.Use(httprate.LimitBy(requestLimit, windowLength, httprate.KeyFromContext(middleware.GetClientIP)))
-func LimitByRealIP(requestLimit int, windowLength time.Duration) func(next http.Handler) http.Handler {
-	return Limit(requestLimit, windowLength, WithKeyFuncs(KeyByRealIP))
 }
 
 func Key(key string) func(r *http.Request) (string, error) {
@@ -105,43 +79,6 @@ func KeyFromContext(extractor func(ctx context.Context) string) KeyFunc {
 	}
 }
 
-// Deprecated: KeyByRealIP trusts the client-supplied True-Client-IP,
-// X-Real-IP, and X-Forwarded-For headers without verifying any proxy chain, so
-// a remote attacker can forge the rate-limit key — see GHSA-9g5q-2w5x-hmxf,
-// GHSA-rjr7-jggh-pgcp, GHSA-3fxj-6jh8-hvhx for the equivalent flaw in chi's
-// middleware.RealIP. On a rate-limiter this is two-sided: an attacker can evade
-// the limit by rotating the spoofed header (unbounded buckets) or lock a victim
-// out by pinning the header to the victim's IP (exhausting their bucket).
-//
-// Install one of chi's middleware.ClientIPFrom* middlewares (chi v5.3.0+) and
-// use KeyFromContext(middleware.GetClientIP) instead:
-//
-//	r.Use(middleware.ClientIPFromXFF("10.0.0.0/8"))
-//	r.Use(httprate.LimitBy(100, time.Minute, httprate.KeyFromContext(middleware.GetClientIP)))
-func KeyByRealIP(r *http.Request) (string, error) {
-	var ip string
-
-	if tcip := r.Header.Get("True-Client-IP"); tcip != "" {
-		ip = tcip
-	} else if xrip := r.Header.Get("X-Real-IP"); xrip != "" {
-		ip = xrip
-	} else if xff := r.Header.Get("X-Forwarded-For"); xff != "" {
-		i := strings.Index(xff, ", ")
-		if i == -1 {
-			i = len(xff)
-		}
-		ip = xff[:i]
-	} else {
-		var err error
-		ip, _, err = net.SplitHostPort(r.RemoteAddr)
-		if err != nil {
-			ip = r.RemoteAddr
-		}
-	}
-
-	return canonicalizeIP(ip), nil
-}
-
 func KeyByEndpoint(r *http.Request) (string, error) {
 	return r.URL.Path, nil
 }
@@ -156,16 +93,6 @@ func WithKeyFuncs(keyFuncs ...KeyFunc) Option {
 
 func WithKeyByIP() Option {
 	return WithKeyFuncs(KeyByIP)
-}
-
-// Deprecated: WithKeyByRealIP installs the spoofable KeyByRealIP and lets a
-// remote attacker forge the rate-limit key — see GHSA-9g5q-2w5x-hmxf,
-// GHSA-rjr7-jggh-pgcp, GHSA-3fxj-6jh8-hvhx for the equivalent flaw in chi's
-// middleware.RealIP. Install one of chi's middleware.ClientIPFrom* middlewares
-// (chi v5.3.0+) and use LimitBy with KeyFromContext(middleware.GetClientIP)
-// instead.
-func WithKeyByRealIP() Option {
-	return WithKeyFuncs(KeyByRealIP)
 }
 
 func WithLimitHandler(h http.HandlerFunc) Option {
